[DEBT-043-D16] Split api/models/portfolio.py into focused modules
- Split api/models/portfolio.py (428 LoC) into 6 files:
  - _balance.py (20): PortfolioBalance model
  - _position.py (47): PortfolioPosition model
  - _fill.py (81): Fill, FillPage models
  - _settlement.py (66): Settlement, SettlementPage models
  - _order.py (234): Order, OrderPage, and all order response models
  - portfolio.py (55): Re-export hub preserving backwards-compatible imports
- Models grouped by domain (balance, position, fill, settlement, order)
- All public imports preserved via re-exports
- Quality gates passed: pre-commit, mypy, pytest (1052 passed)
- Acceptance criteria: 1/1 complete
